Module Swhid_core

module Compute : sig ... end

Module to compute a swhid for all kinds of objects.

module Object : sig ... end

This module contains the various types used to represent a Software Heritage persistent identifier (swhid). These identifiers are documented on Software Heritage.