Module Swhid_core.Compute

Module to compute a swhid for all kinds of objects.

type directory_entry_kind =

| File

| Dir

The type of directory entries.

module Make (SHA1 : sig ... end) (OS : sig ... end) : sig ... end

A functor that, given a SHA1 module and a OS module, provides various functions to compute the swhid of a given object. Supported objects are content, directory, release, revision and snapshot. The origins and visits objects are not supported.