Parameter Make.OS

A module containing the needed OS-related functions.

val contents : string -> string list option

contents dir returns the list of files in the directory dir.

val typ : string -> directory_entry_kind option

type file returns Dir if file is a directory and File otherwise.

val read_file : string -> string option

read_file f returns the content of the file f.

val permissions : string -> int option

permissions f returns the 16-bit file mode (as stored by Git) of the file f. That is:

0o120000 if f is a symlink

0o040000 if f is a directory

0o100755 if f is an executable file

0o100644 if f is a regular file

val base : string -> string

base f is the basename of file f.