Module Object.Qualifier

Module to work with qualifiers. They are documented here.

type t =

| Anchor of Core_identifier.t

| Origin of string

| Path of string

| Visit of Core_identifier.t

| Fragment of int * int option

The type of qualifiers.

val of_string : string -> (t, string) result

of_string s is Ok v if s is a valid qualifier, otherwise it is Error e.

val pp : Format.formatter -> t -> unit

pp fmt v prints v on formatter fmt.

val to_string : t -> string

to_string v is the representation of v as a string.